MiniReaderAlg.h, MiniReaderJets.h, MiniReaderAlg.cxx, MiniReaderJet.cxx: Switch  to the new histogram interface.
MiniReaderJets.cxx:
Testing the HistoBaseClass on jets, added InitJetHisto() and FillJetHisto()
functions in the MiniReaderJets class and calling them in the MiniReaderAlg.

MiniReaderAlg.cxx:
Added a call to the new functions in the MiniReaderJets class.

MiniReaderAlg.h:
Added a typedef for a tuple to be passed as argument to the Fill functions to
allow cross class variables for the plots.


git-svn-id: svn+ssh://svn.cern.ch/reps/atlasinst/Institutes/Stockholm/MonoJetAnalysis/MiniReader_v1/trunk@387489 28d52da6-859c-4cdc-abcc-1ec3d2c02001